pyproject.toml
[project]
name = "binance-trade-bot"
version = "0.1.0"
description = "Binance cryptocurrency trading bot with multiple strategies"
readme = "README.md"
requires-python = ">=3.10,<3.13"
dependencies = [
    "python-binance==1.0.27",
    "sqlalchemy==1.4.15",
    "schedule==1.1.0",
    "apprise==0.9.5.1",
    "Flask==2.3.2",
    "gunicorn==20.1.0",
    "flask-cors==3.0.10",
    "flask-socketio==5.3.6",
    "eventlet==0.30.2",
    "python-socketio[client]==5.11.0",
    "cachetools==4.2.2",
    "sqlitedict==1.7.0",
    "unicorn-binance-websocket-api==1.34.2",
    "unicorn-fy==0.11.0",
    "itsdangerous==2.1.2",
    "Werkzeug==2.3.7",
    "tqdm>=4.67.1",
    "numpy>=2.2.6",
]
